failed attempt to fix memleak
